Update before merging the OOP_preconditioner to master. The work on OOP_preconditioner is not complete, but it's limited to a new GUI to choose the best preconditioner to use in the simulation

+++ oopBoussinesqEquation/src/org/boussinesq/song/Song.java
@@ -13,12 +13,12 @@ public class Song {
 	int xLength;
 	// double[] porosity;
 	double hydraulicConductivity;
-	int t;
+	double t;
 	int alpha;
 	static int h1;
 	int nmax;
 
-	public Song(int time, int Np, double ks) {
+	public Song(double time, int Np, double ks) {
 
 		x = new double[Np];
 		xLength = x.length;
